Indices

Introduction

Indices attempt to quantitatively assess the reliability of a system using a single numerical value.

Expected Energy Not Served (EENS) [MWh]

This is the amount of electricity demand of a user, measured in MWh, that is expected not to be met in a given year.

MultiStateSystems.EENS — Method

EENS(usr::MultiStateSystems.PropDict)

Expected Energy Not Served (EENS) [MWh]

EENS(usr) gives the EENS when an user's demand equals the maximal output of the system for that user.

Generation Ratio Availability (GRA) [-]

This is the probability of at least transferring a specific percentage of the generation to a user through the network.

MultiStateSystems.GRA — Method

GRA(usr::MultiStateSystems.PropDict)

Generation Ratio Availability (GRA) [-]

GRA(usr,GR) gives the probability of transferring at least the GR to a user through the network.

GRA(usr) gives a user's GRA for a GR ranging from 0.0 to 1.0.

GRO(usr,GR) gives the output towards a user for a given GR