Universal Generating Operator

Introduction

Solving a network ntw through the univeral generating operator may be accomplished through:

MultiStateSystems.solve! — Method

solve!(ntw::MultiStateSystems.AbstractNetwork)

This function determines the universal generating function related to the output of all users usr of a network ntw.